+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Test: Improve GameAI tests with configurable options matrix
 * Test: Changed GameAI TreeStrap test to verify best move is maintained after training
 * Viewer: Fix issue when using newgame via the Engine Console not updating the GameType metadata
+* Viewer: Improve BoardHiustory UI and Review Mode performance
 * Viewer: Sort options in Engine Options dialog
 * Viewer: Engine console now scrolls to keep up with the latest output